gateway/requirements.txt
# ============================================================
# Django Gateway - Dependencias de Producción
# Servicio: gateway/ (Puerto 8001 - Gateway principal del sistema)
# ============================================================

# --- Django Core ---
django==4.2.16
djangorestframework==3.15.2
drf-spectacular==0.27.2

# --- Configuración y entorno ---
django-environ==0.11.2
python-dotenv==1.1.1

# --- Base de datos ---
psycopg2-binary==2.9.10
dj-database-url==2.2.0

# --- HTTP Clients ---
requests==2.32.5
httpx==0.28.1

# --- Tareas asíncronas (Celery) ---
celery==5.4.0
redis==5.2.1
django-celery-results==2.5.1
django-celery-beat==2.6.0

# --- Templates / Emails ---
jinja2==3.1.4
resend==2.29.0

# --- Pagos ---
mercadopago==2.2.2

# --- Seguridad ---
cryptography==44.0.0

# --- Deploy / WSGI ---
gunicorn==22.0.0
